backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
bcrypt==5.0.0
cffi==2.0.0
click==8.3.0
cryptography==46.0.3
python-dotenv==1.0.1
pydantic-settings==2.5.2
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup==1.3.0
fastapi==0.119.0
greenlet==3.2.4
h11==0.16.0
idna==3.11
passlib==1.7.4
psycopg2-binary==2.9.11
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.2
pydantic_core==2.41.4
python-jose==3.5.0
python-multipart==0.0.9
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.37.0
web3==6.15.1
eth-account==0.10.0
